+++ nf_core/__main__.py
@@ -520,8 +520,9 @@ def remove(ctx, dir, tool):
 @click.option("-n", "--no-meta", is_flag=True, default=False, help="Don't use meta map for sample information")
 @click.option("-f", "--force", is_flag=True, default=False, help="Overwrite any files if they already exist")
 @click.option("-c", "--conda-name", type=str, default=None, help="Name of the conda package to use")
+@click.option("-p", "--conda-package-version", type=str, default=None, help="Version of conda package to use")
 @click.option("-r", "--repo-type", type=click.Choice(["pipeline", "modules"]), default=None, help="Type of repository")
-def create_module(ctx, tool, dir, author, label, meta, no_meta, force, conda_name, repo_type):
+def create_module(ctx, tool, dir, author, label, meta, no_meta, force, conda_name, conda_package_version, repo_type):
     """
     Create a new DSL2 module from the nf-core template.
 
@@ -542,7 +543,9 @@ def create_module(ctx, tool, dir, author, label, meta, no_meta, force, conda_nam
 
     # Run function
     try:
-        module_create = nf_core.modules.ModuleCreate(dir, tool, author, label, has_meta, force, conda_name, repo_type)
+        module_create = nf_core.modules.ModuleCreate(
+            dir, tool, author, label, has_meta, force, conda_name, conda_package_version, repo_type
+        )
         module_create.create()
     except UserWarning as e:
         log.critical(e)